Replace `nslog` with plain text log
