--- src/winston/nslog.js
@@ -1,16 +1,0 @@
-import nslog from 'nslog'
-import moment from 'moment'
-import _ from 'lodash'
-
-export default class Nslog {
-  constructor () {
-    this.name = 'nslog'
-  }
-
-  log (level, msg, meta, callback) {
-    var output = _.extend({ level: level, message: msg, timestamp: moment() }, meta)
-    nslog(JSON.stringify(output))
-
-    callback(null, true)
-  }
-}
